pyproject.toml
[project]
name = "data-science"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "google-adk>=1.5.0",
    "immutabledict>=4.2.1",
    "sqlglot>=26.10.1",
    "db-dtypes>=1.4.2",
    "regex>=2024.11.6",
    "tabulate>=0.9.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "absl-py>=2.2.2",
    "pydantic>=2.11.3",
    "pandas>=2.0.0,<2.3.0",
    "numpy>=1.25,<2.3.0",
    "pg8000>=1.31.2",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.4.0",
    "joblib>=1.3.0",
    "statsmodels>=0.14.0",
    "autogluon.tabular>=1.4.0",
    "autogluon.multimodal>=1.4.0",
    "autogluon.timeseries>=1.4.0",
    "langchain-text-splitters>=0.3.11",
    "litellm>=1.55.10",
    "openai>=1.59.7",
    "shap>=0.45.0",
    "reportlab>=4.0.0",
    "pillow>=10.0.0",
    "nest-asyncio>=1.6.0",
]


[project.optional-dependencies]
dev = [
    "google-cloud-aiplatform[adk,agent-engines,evaluation]>=1.93.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "google-adk[eval]>=1.5.0",
]


[build-system]
requires = ["uv_build>=0.7.19,<0.8.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "data_science"
module-root = ""

[tool.pytest.ini_options]
# This section is for pytest configuration and does not need to be changed.
console_output_style = "progress"
addopts = "-vv -s"
#addopts = "-vv -s --pdb"
testpaths = ["tests/"]
log_level = "DEBUG"
log_cli = true
log_auto_indent = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = [
    #"error",
    "ignore::UserWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    #'ignore:function ham\(\) is deprecated:DeprecationWarning',
]
